requirements.txt
faiss-cpu
sentence-transformers
pdfplumber  # <--- NEW
fastapi
uvicorn
requests
streamlit
numpy
pydantic